Fixed a minor bug related to flag save/restore and function calls. Still have a bug with the watchpoints tool when using it with the whole_func tool.

+++ granary/arch/x86-64/assemble/5_save_and_restore_flags.cc
@@ -105,7 +105,7 @@ void InjectRestoreFlags(Fragment *frag) {
     }
 
     if (flags.s.of) {
-      APP(ADD_GPR8_IMMb_80r0(&ni, XED_REG_AL, 0x7F));
+      APP(ADD_GPR8_IMMb_80r0(&ni, XED_REG_AL, static_cast<uint8_t>(0x7F)));
     }
 
     APP(SAHF(&ni));
